requirements.txt
Flask==2.2.2
jdtt==1.0.9
waitress==2.1.2
mysql-connector-python==8.0.27
